logging changes and modified amux to include logger and ID. basic amux test change
