+++ mixed_mode_simulator/amux.py
@@ -1,10 +1,10 @@
 import simpy
 from .events import *
 from .sample_and_hold import AnalogBuffer
-
+from EventLogger import EventLogger
 
 class AMUX:
-    def __init__(self, env: simpy.Environment, channels: int, downstream_buffers: list, amux_delay=0, debug=False):
+    def __init__(self, env: simpy.Environment, channels: int, downstream_buffers: list, logger: EventLogger, unitID = 0, amux_delay=0, debug=False):
         """
         Analog MUX simulation model. Takes in a list of downstream buffers and stores them in a Simpy Store FIFO.
         Downstream buffers are allocated until they are present in the FIFO. If FIFO is empty the entire event is
@@ -20,6 +20,7 @@ class AMUX:
         :param debug: Debug flag, default False
         """
         self.amux_delay = amux_delay
+        self.unitID = unitID
         self.env = env
         self.channels = channels
         self.active_channels = {}  # dict of type {channel_index: buffer_resource}
@@ -27,7 +28,7 @@ class AMUX:
         self.downstream_buffers = downstream_buffers
         self.downstream_buffer_fifo = simpy.Store(self.env, capacity=self.num_downstream_buffers)
         self.debug = debug
-
+        self.logger = logger
         self.fill_fifo(self.downstream_buffers)
 
     def fill_fifo(self, buffers: list[AnalogBuffer]):
@@ -65,6 +66,8 @@ class AMUX:
         event.event.fail(Exception(f'Event {event.detection_event_info["event_number"]}, sample number '
                   f'{event.event_info["sample_index"]} '
                   f'dropped. No downstream channels available'))
+        #self.logger.log_event()
+        
 
     def entry_point(self, channel_index: int, event: DownstreamEvent):
 
@@ -84,10 +87,21 @@ class AMUX:
             if event.final_event:
                 if self.debug:
                     print(f'Downstream buffer {self.active_channels[channel_index]["buffer"].buffer_index} released at {self.env.now}')
+                    yield self.env.process(self.logger.log_event(
+                    'amux',  # Component name
+                    self.unitID,  # Unit index, now using the unique identifier
+                    event
+                    ))
                 self.release_buffer(channel_index)
+                
 
         else:
             """
             Not in active channel. Event is dropped. 
             """
             self.drop_event(event)
+            yield self.env.process(self.logger.log_event(
+            'amux',  # Component name
+            self.unitID,  # Unit index, now using the unique identifier
+            event
+        ))
