+++ tests/analog_buffer_test.py
@@ -1,6 +1,6 @@
 import unittest
-from ..mixed_mode_simulator import sample_and_hold
-from ..mixed_mode_simulator import events
+from mixed_mode_simulator import sample_and_hold
+from mixed_mode_simulator import events
 import simpy
